api/requirements.txt
appdirs==1.4.4
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
bidict==0.21.2
black==20.8b1
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
dataclasses==0.8
fastapi==0.63.0
h11==0.12.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.8.1
pycparser==2.20
pydantic==1.8.1
PyJWT==2.0.1
PyMySQL==1.0.2
regex==2020.11.13
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
websockets==8.1
